+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   `reaper-entry-is-running`, `reaper-entry-hours` and
   `reaper-entry-notes` macros to handle time entries.
 - `reaper-with-selected-timer` for working with the timer at point.
+- `reaper-get-project`, `reaper-get-head-project`,
+  `reaper-project-id`, `reaper-project-code`, `reaper-project-name`
+  and `reaper-project-tasks` for working with projects.
 
 ### Changed
 - Use `(goto-char (point-min))` instead of `(beginning-of-buffer)`